refactor pages loadContent
